Move caching-related functionality to a separate module
